refactor(testing): split fee payment tests per auth component
Per review: the fee payment tests now live in a module directory with
one file per auth flavor (singlesig, network), sharing the chain
constants via the module root.
