+++ crates/miden-testing/tests/auth/fee_payment/mod.rs
@@ -1,0 +1,9 @@
+mod network;
+mod singlesig;
+
+// CONSTANTS
+// ================================================================================================
+
+/// The verification base fee configured on the fee-charging mock chains used across the fee
+/// payment tests.
+const VERIFICATION_BASE_FEE: u32 = 500;
